requirements.txt
gradio==4.19.2
langchain==0.1.7
langchain-openai==0.0.5
minio==7.2.0
paramiko==3.3.1
python-dotenv==1.0.0
openai==1.12.0